% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_BIP.R
\name{get_BIP}
\alias{get_BIP}
\title{Retrieve BIP scores for a list of DOIs}
\usage{
get_BIP(dois, verbose = FALSE)
}
\arguments{
\item{dois}{A character vector of DOIs}

\item{verbose}{Show diagnostic information?}
}
\value{
A data frame with bibliometric indicators for the provided DOIs
}
\description{
This function takes a list of DOIs, normalizes them, queries the BIP API to retrieve scores, and returns a data frame with bibliometric indicators.
}
\examples{
dois <- c("https://doi.org/10.1080/00223891.2020.1726936", "10.1016/j.jrp.2013.05.009")
get_BIP(dois)
}
