% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_set.R
\name{get_reference_set_from_files}
\alias{get_reference_set_from_files}
\title{Helper function: Get reference set from files}
\usage{
get_reference_set_from_files(path)
}
\arguments{
\item{path}{The relative or absolute path of the files}
}
\value{
Returns a data frame with a reference set.
}
\description{
The \code{get_reference_set} function can save intermediate files on a drive. In case that the download or the function aborts with an error, all existing files in a given path can be read combined.
}
