% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_index.R
\name{h_index}
\alias{h_index}
\title{Compute h-index and hIa-index}
\usage{
h_index(
  search = NULL,
  display_name = NULL,
  author.id = NULL,
  first_pub_year = NA,
  academic_age_bonus = 0
)
}
\arguments{
\item{search}{A character string with the author's name for searching in the OpenAlex API. Allows fuzzy search (e.g. with or without middle initial).}

\item{display_name}{A character vector with the exact author's names for searching in the OpenAlex API. You cvan provide multiple variants in a vector.}

\item{author.id}{A character vector with OpenAlex author ID(s) (i.e., an ID starting with "A.......") for searching in the OpenAlex API.}

\item{first_pub_year}{An integer with the year of the author's first publication. If not provided, it is estimated from the retrieved publications.}

\item{academic_age_bonus}{Years that are subtracted from active academic life (e.g. due to child care). For example, SNF subtracts 1.5 years per child (see https://www.snf.ch/en/cciM9NWuvhOVKRxv/news/news-200803-career-funding-the-researchers-overall-performance-counts).}
}
\value{
A list containing the net academic age, h-index and hIa-index of the author.
}
\description{
Computes the h-index and hIa-index of an author. You need to provide exactly *one* of the parameters `search`, `display_name` or `author.id`:
- `search` allows fuzzy search of the author's name (e.g. with or without middle initial). Might return too many matches
- `display_name` allows to provide a vector with multiple versions of a name (e.g., `c("Hans Müller", "Hans Mueller", "Hans J. Muller")`)
- `author.id` is the OpenAlex ID for authors.
}
\examples{
# Compute the h-index and hIa-index of Albert Einstein
h_index(search = "Felix Schönbrodt")
h_index("Markus Bühner", first_pub_year = 1997) # adjust year of first publication

# If you have alternative names, you need to use the `display_name` parameter
# with exact spelling (all variants that are relevant)
h_index(display_name = c("Ramona Schödel", "Ramona Schoedel"))

}
