% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{normalize_dois}
\alias{normalize_dois}
\title{Normalize DOIs}
\usage{
normalize_dois(x, verbose = FALSE)
}
\arguments{
\item{x}{Character vector of DOIs to be normalized}
}
\value{
Character vector of normalized DOIs
}
\description{
This function normalizes DOIs by trimming leading/trailing whitespace,
replacing 'dx.doi.org' with 'doi.org', replacing 'http:' with 'https:',
and converting the DOI to lowercase, according to the DOI specification
}
\examples{
normalize_dois(c("  10.123.10.1/DOI ", "http://dx.doi.org/10.456/Doi"))
}
