% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{normalize_ORCIDs}
\alias{normalize_ORCIDs}
\title{Normalize ORCIDs}
\usage{
normalize_ORCIDs(x, verbose = FALSE)
}
\arguments{
\item{x}{Character vector of ORCIDs to be normalized}

\item{verbose}{Show diagnostic information?}
}
\value{
Character vector of normalized ORCIDS
}
\description{
This function normalizes DOIs by trimming leading/trailing whitespace,
replacing 'dx.doi.org' with 'doi.org', replacing 'http:' with 'https:',
and converting the DOI to lowercase, according to the DOI specification
}
\examples{
normalize_ORCIDs(c("orcid.org/0000-1234-5678-9111", "0000-1234-5678-9111", "  https://orcid.org/0000-1234-5678-9111 "))
}
