% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FNCR.R
\name{upper_trim_mean}
\alias{upper_trim_mean}
\title{Calculates the upper trimmed mean of a vector}
\usage{
upper_trim_mean(x, trim)
}
\arguments{
\item{x}{A numeric vector}

\item{trim}{A numeric value between 0 and 1 indicating the proportion of values to trim from the top}
}
\value{
A numeric value representing the upper trimmed mean of the vector
}
\description{
Calculates the upper trimmed mean of a vector
}
\examples{
x <- c(1,2,3,4,5,6,7)
upper_trim_mean(x, 0.2)

}
