% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FNCR.R
\name{ecdf2}
\alias{ecdf2}
\title{ecdf2}
\usage{
ecdf2(x)
}
\arguments{
\item{x}{A vector of values.}
}
\value{
A function that calculates the linearly interpolated ecdf of the vector of values.
}
\description{
Calculates an linearly interpolated empirical cumulative distribution function (ecdf) from a vector of values. This function was provided by Tal Galili on \href{https://stats.stackexchange.com/q/230458}{StackOverflow}.
}
\examples{
x <- c(1,2,3,4,5,6,7,8)
ecdf2(x)
}
