% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FNCS.R
\name{FNCS}
\alias{FNCS}
\title{Compute the Field Normalized Citation Score (FNCS) and Percentage Rank (FNPR) of a publication}
\usage{
FNCS(dois, ref_set, upper_trim = 0)
}
\arguments{
\item{dois}{A character vector of the DOI of the paper for which the FNCS should be computed.}

\item{ref_set}{A data frame containing the reference set for the paper of interest. This is an object from the \code{get_reference_set} function.}

\item{upper_trim}{A numeric value between 0 and 1 that indicates the fraction of values to be trimmed from the upper end of the reference set. Scheidsteger et al. (2023) remove the upper 1 percent of citation counts when using OpenAlex. This only affects the FNCS, not the percentile rank (FNPR).}
}
\value{
A list containing the computed FNCS and the percentile rank of the paper. The latter is the CP-EX measure which means "how many citations in the reference set have \emph{less} citations than the target paper".
}
\description{
The FNCS is defined as ... (TODO)
The percentage rank is the "CP-EX" measure described in Bornmann & Williams (2020), which returns the percentage of publication with less citations (and "less or equal"). As many publications have 0 citations, the alternative CP-IN measure (which return "less or equal") returns often very high percentiles although a paper has 0 citations. This is not intuitive. For the percentile rank, the function uses a linear interpolation (cf. Bornmann & Williams, 2020) using a function provided by Tal Galili (https://stats.stackexchange.com/q/230458)
}
\examples{
\dontrun{
# TODO: ref_set does not exist yet
FNCS(dois = "10.1038/s41586-019-1712-z", ref_set = ref_set, upper_trim = .01)
}
}
\references{
Bornmann, L., & Williams, R. (2020). An evaluation of percentile measures of citation impact, and a proposal for making them better. Scientometrics, 124(2), 1457–1478. https://doi.org/10.1007/s11192-020-03512-7
Scheidsteger, T., Haunschild, R., & Bornmann, L. (2023). How similar are field-normalized scores from different free or commercial databases calculated for large German universities? https://dapp.orvium.io/deposits/6441118c643beb0d90fc543f/view
}
