% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FNCS.R
\name{FNCS}
\alias{FNCS}
\title{Compute the Field Normalized Citation Score of a publication}
\usage{
FNCS(doi, ref_set, upper_trim = 0.01)
}
\arguments{
\item{doi}{A character string of the DOI of the paper for which the FNCS should be computed.}

\item{ref_set}{A data frame containing the reference set for the paper of interest. This is an object from the \code{get_reference_set} function.}

\item{upper_trim}{A numeric value between 0 and 1 that indicates the fraction of values to be trimmed from the upper end of the reference set. Scheidsteger et al. (2023) remove the upper 1 percent of citation counts when using OpenAlex. This only affects the FNCS, not the percentile rank.}
}
\value{
A list containing the computed FNCS and the percentile rank of the paper.
}
\description{
The percentage rank is the "CP-IN" measure described in Bornmann & Williams (2020). For the percentile rank, the function uses a linear interpolation (cf. Bornmann & Williams, 2020) using a function provided by Tal Galili (https://stats.stackexchange.com/q/230458)
}
\examples{
\dontrun{
# TODO: ref_set does not exist yet
FNCS(doi = "10.1038/s41586-019-1712-z", ref_set = ref_set, upper_trim = .01)
}
}
\references{
Bornmann, L., & Williams, R. (2020). An evaluation of percentile measures of citation impact, and a proposal for making them better. Scientometrics, 124(2), 1457–1478. https://doi.org/10.1007/s11192-020-03512-7
Scheidsteger, T., Haunschild, R., & Bornmann, L. (2023). How similar are field-normalized scores from different free or commercial databases calculated for large German universities? https://dapp.orvium.io/deposits/6441118c643beb0d90fc543f/view
}
