% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JIF.R
\name{get_JIF}
\alias{get_JIF}
\title{Compute the Journal Impact Factor}
\usage{
get_JIF(issn, year, verbose = FALSE)
}
\arguments{
\item{issn}{The ISSN of the journal}

\item{year}{The target year}

\item{verbose}{Whether to print verbose output (default is FALSE)}
}
\value{
A data frame with columns for journal, ISSN, year, total_citations,
citable_items and JIF.
}
\description{
This function computes the Journal Impact Factor (JIF) for a given journal
given its ISSN and a target year.
}
\examples{
\dontrun{
get_JIF("0022-3514", 2018)
}

}
