% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_index.R
\name{h_index}
\alias{h_index}
\title{Compute h-index and hIa-index}
\usage{
h_index(
  search = NULL,
  display_name = NULL,
  author.id = NULL,
  first_pub_year = NA,
  academic_age_bonus = 0
)
}
\arguments{
\item{search}{A character string with the author's name for searching in the OpenAlex API. Allows fuzzy search (e.g. with or without middle initial).}

\item{display_name}{A character vector with the exact author's names for searching in the OpenAlex API. You cvan provide multiple variants in a vector.}

\item{author.id}{A character vector with OpenAlex author ID(s) (i.e., an ID starting with "A.......") for searching in the OpenAlex API.}

\item{first_pub_year}{An integer with the year of the author's first publication. If not provided, it is estimated from the retrieved publications.}

\item{academic_age_bonus}{Years that are subtracted from active academic life (e.g. due to child care). For example, SNF subtracts 1.5 years per child (see https://www.snf.ch/en/cciM9NWuvhOVKRxv/news/news-200803-career-funding-the-researchers-overall-performance-counts).}
}
\value{
A list containing the net academic age, h-index and hIa-index of the author.
}
\description{
TODO: Could use a time-window (currently, it uses all citations since publication).
}
\details{
Computes the h-index and hIa-index of an author. You need to provide exactly \emph{one} of the parameters \code{search}, \code{display_name} or \code{author.id}:
\itemize{
\item \code{search} allows fuzzy search of the author's name (e.g. with or without middle initial). Might return too many matches
\item \code{display_name} requires an exact match but allows to provide a vector with multiple versions of a name (e.g., \code{c("Hans Müller", "Hans Mueller", "Hans J. Muller")})
\item \code{author.id} is the OpenAlex ID for authors.
}

Often the same author is listed under different IDs in the OpenAlex database, therefore these are merged by the function. \strong{But what if the function includes wrong authors?} Then you need to manually screen the table with author names, record the IDs from the correct aliases of the target person (starting with "A.......") and enter these author IDs into the \code{author.id} parameter.
}
\examples{
h_index(search = "Felix Schönbrodt")
h_index("Markus Bühner", first_pub_year = 1997) # adjust year of first publication


}
