% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{normalize_ORCIDs}
\alias{normalize_ORCIDs}
\title{Normalize ORCIDs}
\usage{
normalize_ORCIDs(x)
}
\arguments{
\item{x}{Character vector of ORCIDs to be normalized}
}
\value{
Character vector of normalized ORCIDS
}
\description{
This function normalizes ORCIDs to the canonical form
(which is the URL to the profile)
}
\examples{
normalize_ORCIDs(c("orcid.org/0000-1234-5678-9111",
    "0000-1234-5678-9111", "  https://orcid.org/0000-1234-5678-9111 "))
}
